new_testament/requirements.txt
# New Testament: St. Stella's Genomic Analysis Framework
# Core dependencies for high-performance genomic coordinate transformation
# Python 3.13 compatible versions

# Core scientific computing - Python 3.13 compatible
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0

# High-performance computing and parallelization
# Note: numba is optional for Python 3.13 - framework provides NumPy fallback
# Uncomment the following line if you want high-performance JIT compilation:
# numba>=0.58.0,<1.0.0

# Data manipulation and analysis
pandas>=2.0.0,<3.0.0

# Visualization and plotting
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Bioinformatics core
biopython>=1.81,<2.0.0

# System monitoring and performance
psutil>=5.9.0,<6.0.0

# Configuration and data I/O
pyyaml>=6.0,<7.0.0
h5py>=3.8.0,<4.0.0

# Testing and validation
pytest>=7.0.0,<8.0.0
pytest-benchmark>=4.0.0,<5.0.0

# Progress tracking and user interface
tqdm>=4.64.0,<5.0.0

# Parallel processing utilities
joblib>=1.3.0,<2.0.0

# Memory profiling and optimization
memory-profiler>=0.61.0,<1.0.0

# Optional but recommended for advanced features:
# GPU acceleration (uncomment if CUDA available):
# cupy>=10.0.0,<12.0.0

# Enhanced visualization (uncomment if needed):
# plotly>=5.0.0,<6.0.0
# bokeh>=2.4.0,<3.0.0
# networkx>=2.6.0,<3.0.0

# Machine learning integration (uncomment if needed):
# scikit-learn>=1.0.0,<2.0.0
# tensorflow>=2.8.0,<3.0.0
# torch>=1.11.0,<3.0.0

# Advanced bioinformatics (uncomment if working with BAM/VCF files):
# pysam>=0.19.0,<1.0.0
# pyvcf>=0.6.8,<1.0.0
# pyranges>=0.0.120,<1.0.0

# Development dependencies (uncomment for development):
# black>=22.0.0,<24.0.0
# flake8>=4.0.0,<7.0.0
# mypy>=0.950,<2.0.0
# pre-commit>=2.17.0,<4.0.0
# pytest-cov>=3.0.0,<5.0.0
